ChangeActivityState for MultiInstance and Timer (when changing state from a task with a timer) Separate related changeActivityState tests from RuntimeServiceTests into RuntimeServiceChangeStateTest
