+++ modules/flowable-engine/src/main/java/org/flowable/engine/impl/persistence/entity/ExecutionEntity.java
@@ -13,6 +13,9 @@
 
 package org.flowable.engine.impl.persistence.entity;
 
+import static java.util.Comparator.comparing;
+
+import java.util.Comparator;
 import java.util.Date;
 import java.util.List;
 
@@ -36,6 +39,8 @@ import org.flowable.task.service.impl.persistence.entity.TaskEntity;
  */
 
 public interface ExecutionEntity extends DelegateExecution, Execution, ProcessInstance, Entity, AlwaysUpdatedPersistentObject, HasRevision {
+
+    Comparator<ExecutionEntity> EXECUTION_ENTITY_START_TIME_ASC_COMPARATOR = comparing(ProcessInstance::getStartTime);
 
     void setBusinessKey(String businessKey);
