+++ modules/flowable-engine/src/test/resources/org/flowable/engine/test/api/oneTaskWithTimerInSubProcess.bpmn20.xml
@@ -1,51 +1,50 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<definitions id="definitions" xmlns="http://www.omg.org/spec/BPMN/20100524/MODEL"
-  xmlns:activiti="http://activiti.org/bpmn" targetNamespace="Examples">
+<definitions id="definitions" xmlns="http://www.omg.org/spec/BPMN/20100524/MODEL" targetNamespace="Examples">
 
-  <process id="startSimpleSubProcess">
+    <process id="startSimpleSubProcess">
 
-    <startEvent id="theStart" />
+        <startEvent id="theStart"/>
 
-    <sequenceFlow id="flow1" sourceRef="theStart" targetRef="taskBefore" />
-    
-    <userTask id="taskBefore" name="Task before sub process" />
-    
-    <sequenceFlow id="flow2" sourceRef="taskBefore" targetRef="subProcess" />
+        <sequenceFlow id="flow1" sourceRef="theStart" targetRef="taskBefore"/>
 
-    <subProcess id="subProcess">
+        <userTask id="taskBefore" name="Task before sub process"/>
 
-      <startEvent id="theSubProcessStart" />
+        <sequenceFlow id="flow2" sourceRef="taskBefore" targetRef="subProcess"/>
 
-      <sequenceFlow id="subflow1" sourceRef="theSubProcessStart" targetRef="subtask" />
+        <subProcess id="subProcess">
 
-      <userTask id="subtask" name="Task in subprocess" />
-      
-      <boundaryEvent id="boundaryTimerEvent" attachedToRef="subtask">
-        <timerEventDefinition>
-          <timeDuration>PT1H</timeDuration>
-        </timerEventDefinition>
-      </boundaryEvent>
-      
-      <sequenceFlow id="subflow2" sourceRef="boundaryTimerEvent" targetRef="subtask2" />
-      
-      <userTask id="subtask2" />
+            <startEvent id="theSubProcessStart"/>
 
-      <sequenceFlow id="subflow3" sourceRef="subtask" targetRef="theSubProcessEnd" />
-      
-      <sequenceFlow id="subflow4" sourceRef="subtask2" targetRef="theSubProcessEnd" />
+            <sequenceFlow id="subflow1" sourceRef="theSubProcessStart" targetRef="subTask"/>
 
-      <endEvent id="theSubProcessEnd" />
+            <userTask id="subTask" name="Task in subProcess"/>
 
-    </subProcess>
+            <boundaryEvent id="boundaryTimerEvent" attachedToRef="subTask">
+                <timerEventDefinition>
+                    <timeDuration>PT1H</timeDuration>
+                </timerEventDefinition>
+            </boundaryEvent>
 
-    <sequenceFlow id="flow4" sourceRef="subProcess" targetRef="taskAfter" />
-    
-    <userTask id="taskAfter" name="Task after sub process" />
-    
-    <sequenceFlow id="flow5" sourceRef="taskAfter" targetRef="theEnd" />
+            <sequenceFlow id="subflow2" sourceRef="boundaryTimerEvent" targetRef="theSubProcessEnd"/>
 
-    <endEvent id="theEnd" />
+            <sequenceFlow id="subflow3" sourceRef="subTask" targetRef="subTask2"/>
 
-  </process>
+            <userTask id="subTask2"/>
+
+            <sequenceFlow id="subflow4" sourceRef="subTask2" targetRef="theSubProcessEnd"/>
+
+            <endEvent id="theSubProcessEnd"/>
+
+        </subProcess>
+
+        <sequenceFlow id="flow4" sourceRef="subProcess" targetRef="taskAfter"/>
+
+        <userTask id="taskAfter" name="Task after sub process"/>
+
+        <sequenceFlow id="flow5" sourceRef="taskAfter" targetRef="theEnd"/>
+
+        <endEvent id="theEnd"/>
+
+    </process>
 
 </definitions>
